update chain transaction for new package
